+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `@vitest/eslint-plugin`: 1.1.28 -> 1.1.31
 - `eslint-import-resolver-typescript`: 3.7.0 -> 3.8.0
 - `eslint-plugin-perfectionist`: 4.8.0 -> 4.9.0
+- `eslint-plugin-sonarjs`: 3.0.1 -> 3.0.2
 
 ## 0.4.2