+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - `eslint-import-resolver-typescript`: 3.7.0 -> 3.8.0
 - `eslint-plugin-perfectionist`: 4.8.0 -> 4.9.0
 - `eslint-plugin-sonarjs`: 3.0.1 -> 3.0.2
+- `eslint-plugin-yml`: 1.16.0 -> 1.17.0
 
 ## 0.4.2